+++ src/shared/lib/index.ts
@@ -26,3 +26,5 @@ export const useDebounce = <F extends (...args: unknown[]) => ReturnType<F>>(
 
   return debouncedValue;
 };
+
+export * from "./cookie";
